refactor: isolate vision classes and switch to java units api
